test: converted tests that don't interact with other contracts to unit tests
jira: DCD-50
