+++ contracts/poap/src/error.rs
@@ -1,7 +1,7 @@
 use cosmwasm_std::{StdError, Timestamp};
 use thiserror::Error;
 
-#[derive(Error, Debug)]
+#[derive(Error, Debug, PartialEq)]
 pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),
